fintech-rag-pilot/requirements.txt
langchain
chromadb
sentence-transformers
streamlit
pypdf2
python-multipart
uvicorn
fastapi
requests
